changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11855 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/BaseClasses/ORDetectorSegment.h
@@ -79,8 +79,8 @@
 - (NSUndoManager*) undoManager;
 - (id) objectForKey:(id)key;
 - (void) setObject:(id)obj forKey:(id)key;
-- (void) setSegmentNumber:(unsigned)index;
-- (unsigned) segmentNumber;
+- (void) setSegmentNumber:(NSUInteger)index;
+- (NSUInteger) segmentNumber;
 - (BOOL) segmentError;
 - (void) setSegmentError:(BOOL)state;
 - (void) clearSegmentError;
